export: change VCS tag in spec to sha1 of commitish
Do not resolve to the sha1 of the commit object anymore. That is, the
VCS tag will contain the sha1 of the tag object if an (annotated) tag is
exported (previously, it would contain the sha1 of the commit the tag,
or the chain of tags, pointed to) (previously, it would contain the sha1
of the commit the tag, or the chain of tags, pointed to).

Change-Id: I5b8d164cf1f4e910740d75df825cd506ffb9437c
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>