Implement 'fallback-to-native' option
Add a new '--fallback-to-native' command line option and a corresponding
'fallback_to_native' configuration option. This option re-introducs the
old behaviour (removed in v0.17) of falling back to to native packaging
mode if upstream tarball or patch generation fails. That is gbs will
simply create the source tarball from the exported commit (instead of
upstream tag or pristine-tar) and doesn't generate any patches.

Change-Id: Iee44872c095fa5e50435e4c57ba17ef27c114aeb
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>