Implement 'native' config option
Add new 'native' configuration file option. The new option explicitly
sets the "nativity" of the package and no guessing based on existing
branches is done.

It is recommended to have this option set in the package-specific (i.e.
committed in the Git tree) .gbs.conf for every package. In the config
file, values 'yes', 'on', '1', 'true' and 'enabled' are interpreted as
True, all other non-empty values are interpreted as False.

Change-Id: Ie7474c04287124fd928e7b8b43d64392dc68ca3c
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>